Application/requirements.txt
Flask
pymongo[srv]
Flask-PyMongo
flask
flask_sessions
python-dotenv